CancellableFuture.repeatWithMode + moving generators to a subpackage
