+++ src/test/scala/io/github/makingthematrix/signals3/generators/GeneratorsSpec.scala
@@ -1,6 +1,9 @@
-package io.github.makingthematrix.signals3
+package io.github.makingthematrix.signals3.generators
 
-import testutils.{awaitAllTasks, waitForResult}
+import io.github.makingthematrix.signals3.generators.{GeneratorEventStream, GeneratorSignal}
+import io.github.makingthematrix.signals3.testutils.{awaitAllTasks, waitForResult}
+import io.github.makingthematrix.signals3.{EventContext, Signal, Threading}
+
 import scala.collection.mutable
 import scala.concurrent.duration.*
